Move partners constant to new file
