+++ app/api/partners/route.tsx
@@ -1,217 +1,5 @@
 import { NextResponse } from 'next/server'
-
-export const partners = [
-  {
-    homepageUrl: 'https://www.vegvesen.no/',
-    logoUrl: 'https://2023.javazone.no/logos/vegvesen.svg',
-    name: 'Statens vegvesen',
-  },
-  {
-    homepageUrl: 'https://www.systek.no/',
-    logoUrl: 'https://2023.javazone.no/logos/systek.svg',
-    name: 'Systek',
-  },
-  {
-    homepageUrl: 'https://www.storebrand.no/',
-    logoUrl: 'https://2023.javazone.no/logos/storebrand.png',
-    name: 'Storebrand',
-  },
-  {
-    homepageUrl: 'https://www.redhat.com/en',
-    logoUrl: 'https://2023.javazone.no/logos/redhat.svg',
-    name: 'Redhat',
-  },
-  {
-    homepageUrl: 'https://www.knowit.no',
-    logoUrl: 'https://2023.javazone.no/logos/knowit.svg',
-    name: 'Knowit Objectnet AS',
-  },
-  {
-    homepageUrl: 'https://www.soprasteria.no/om-sopra-steria',
-    logoUrl: 'https://2023.javazone.no/logos/soprasteria.svg',
-    name: 'Sopra Steria',
-  },
-  {
-    homepageUrl: 'https://www.accenture.com/no-en',
-    logoUrl: 'https://2023.javazone.no/logos/accenture.svg',
-    name: 'Accenture AS',
-  },
-  {
-    homepageUrl: 'https://www.nav.no/no/Person',
-    logoUrl: 'https://2023.javazone.no/logos/nav.svg',
-    name: 'NAV',
-  },
-  {
-    homepageUrl: 'https://www.miles.no/',
-    logoUrl: 'https://2023.javazone.no/logos/miles.svg',
-    name: 'Miles',
-  },
-  {
-    homepageUrl: 'https://www.kantega.no/',
-    logoUrl: 'https://2023.javazone.no/logos/kantega.png',
-    name: 'Kantega AS',
-  },
-  {
-    homepageUrl: 'https://www.fremtind.no/',
-    logoUrl: 'https://2023.javazone.no/logos/fremtind.svg',
-    name: 'Fremtind Forsikring AS',
-  },
-  {
-    homepageUrl: 'https://www.finn.no/',
-    logoUrl: 'https://2023.javazone.no/logos/finn.svg',
-    name: 'FINN.no',
-  },
-  {
-    homepageUrl: 'https://schibsted.com/',
-    logoUrl: 'https://2023.javazone.no/logos/schibsted.png',
-    name: 'Schibsted',
-  },
-  {
-    homepageUrl: 'https://www.dnb.no/',
-    logoUrl: 'https://2023.javazone.no/logos/dnb.svg',
-    name: 'DNB',
-  },
-  {
-    homepageUrl: 'https://www.tietoevry.com/no/',
-    logoUrl: 'https://2023.javazone.no/logos/tietoevry.svg',
-    name: 'TietoEVRY AS',
-  },
-  {
-    homepageUrl: 'https://www.statnett.no/',
-    logoUrl: 'https://2023.javazone.no/logos/statnett.svg',
-    name: 'Statnett SF',
-  },
-  {
-    homepageUrl: 'https://www.jpro.no/',
-    logoUrl: 'https://2023.javazone.no/logos/jpro.svg',
-    name: 'JProfessionals AS',
-  },
-  {
-    homepageUrl: 'https://www.sparebank1.no/',
-    logoUrl: 'https://2023.javazone.no/logos/sparebank1.svg',
-    name: 'SpareBank 1 Banksamarbeidet',
-  },
-  {
-    homepageUrl: 'https://scienta.no/',
-    logoUrl: 'https://2023.javazone.no/logos/scienta.svg',
-    name: 'Scienta',
-  },
-  {
-    homepageUrl: 'https://enso.no/',
-    logoUrl: 'https://2023.javazone.no/logos/enso.svg',
-    name: 'Konsulentselskapet Enso AS',
-  },
-  {
-    homepageUrl: 'https://scelto.no/',
-    logoUrl: 'https://2023.javazone.no/logos/scelto.svg',
-    name: 'Scelto AS',
-  },
-  {
-    homepageUrl: 'https://www.tripletex.no/',
-    logoUrl: 'https://2023.javazone.no/logos/tripletex.svg',
-    name: 'Tripletex',
-  },
-  {
-    homepageUrl: 'https://www.politiet.no/om/organisasjonen/sarorganene/pit/',
-    logoUrl: 'https://2023.javazone.no/logos/politiet.svg',
-    name: 'Politiets IT-enhet',
-  },
-  {
-    homepageUrl: 'https://www.oslo.kommune.no/',
-    logoUrl: 'https://2023.javazone.no/logos/oslo_kommune.svg',
-    name: 'Oslo Origo',
-  },
-  {
-    homepageUrl: 'https://www.webstep.no/',
-    logoUrl: 'https://2023.javazone.no/logos/webstep.svg',
-    name: 'Webstep AS',
-  },
-  {
-    homepageUrl: 'https://bouvet.no/',
-    logoUrl: 'https://2023.javazone.no/logos/bouvet.svg',
-    name: 'Bouvet Norge AS',
-  },
-  {
-    homepageUrl: 'https://computas.com/',
-    logoUrl: 'https://2023.javazone.no/logos/computas.svg',
-    name: 'Computas',
-  },
-  {
-    homepageUrl: 'https://www.gjensidige.no/',
-    logoUrl: 'https://2023.javazone.no/logos/gjensidige.svg',
-    name: 'Gjensidige Forsikring',
-  },
-  {
-    homepageUrl: 'https://www.skatteetaten.no/person/',
-    logoUrl: 'https://2023.javazone.no/logos/skatteetaten.svg',
-    name: 'Skatteetaten',
-  },
-  {
-    homepageUrl: 'https://github.com/tolletaten/javazone2021',
-    logoUrl: 'https://2023.javazone.no/logos/toll.svg',
-    name: 'Tolletaten',
-  },
-  {
-    homepageUrl: 'https://uptimeconsulting.no/',
-    logoUrl: 'https://2023.javazone.no/logos/uptime.svg',
-    name: 'Uptime Consulting',
-  },
-  {
-    homepageUrl: 'https://eika.no/',
-    logoUrl: 'https://2023.javazone.no/logos/eika.svg',
-    name: 'Eika Gruppen AS',
-  },
-  {
-    homepageUrl: 'https://www.capgemini.com/no-no/',
-    logoUrl: 'https://2023.javazone.no/logos/capgemini.svg',
-    name: 'Capgemini Norge AS',
-  },
-  {
-    homepageUrl: 'https://embriq.no',
-    logoUrl: 'https://2023.javazone.no/logos/embriq.svg',
-    name: 'EmbriQ',
-  },
-  {
-    homepageUrl: 'https://bekk.no',
-    logoUrl: 'https://2023.javazone.no/logos/bekk.svg',
-    name: 'Bekk Consulting AS',
-  },
-  {
-    homepageUrl: 'https://cloudberries.no',
-    logoUrl: 'https://2023.javazone.no/logos/cloudberries.svg',
-    name: 'Cloudberries',
-  },
-  {
-    homepageUrl: 'https://ruter.no',
-    logoUrl: 'https://2023.javazone.no/logos/ruter.svg',
-    name: 'Ruter',
-  },
-  {
-    homepageUrl: 'https://mastercard.com',
-    logoUrl: 'https://2023.javazone.no/logos/mastercard.svg',
-    name: 'Mastercard Payment Services Norway AS',
-  },
-  {
-    homepageUrl: 'https://www.kodemaker.no',
-    logoUrl: 'https://2023.javazone.no/logos/kodemaker.svg',
-    name: 'Kodemaker',
-  },
-  {
-    homepageUrl: 'https://www.variant.no',
-    logoUrl: 'https://2023.javazone.no/logos/variant.svg',
-    name: 'Variant',
-  },
-  {
-    homepageUrl: 'https://www.dfindconsulting.no/',
-    logoUrl: 'https://2023.javazone.no/logos/dfind.svg',
-    name: 'Dfind',
-  },
-  {
-    homepageUrl: ' https://www.postennorge.no/',
-    logoUrl: 'https://2023.javazone.no/logos/postenbring.svg',
-    name: 'Posten Bring',
-  },
-]
+import { partners } from './partners'
 
 export async function GET() {
   return NextResponse.json(partners)
